Check if there really is a file to delete before doing unlink.
The (not succeeding) call to unlink might change errno which disturbs
the test code.

+++ src/std/c_file.c
@@ -166,7 +166,7 @@ out:
   if (dstfd >= 0) {
       close(dstfd);
   }
-  if (rc < 0) {
+  if (rc < 0 && c_isfile(dst)) {
     unlink(dst);
   }
   return rc;
